renamed widget and made list tile dismissible
